% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.r
\name{get_top_data}
\alias{get_top_data}
\title{Get TOP factor data}
\usage{
get_top_data(detail = level_detail$minimal)
}
\arguments{
\item{detail}{of detail of returned tibble, NULL for all columns}
}
\value{
tibble
}
\description{
Returns the full TOP factor dataset
as a tibble
}
