% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_updated.r
\name{last_updated}
\alias{last_updated}
\title{When was the TOP data last updated?}
\usage{
last_updated()
}
\value{
a POSIX formatted date
}
\description{
Access the open science framework API and
check when the TOP data was last modified
}
