% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topapi-package.r
\docType{package}
\name{topapi}
\alias{topapi-package}
\title{topapi: Tools for handling the TOP dataset}
\description{
Helper functions for working with the TOP dataset.
}
\details{
The \strong{topapi} package implements helper functions for
smooth querying of the TOP factor dataset.

Rather than manually downloading the dataset,
the \strong{topapi} package keeps the dataset in memory,
meaning the data is both easily accessible and
always up to date.

Main exported methods are:
\itemize{
\item Data retrieval: \code{\link[=journal_name]{journal_name()}}, \code{\link[=journal_issn]{journal_issn()}}, \code{\link[=publisher_name]{publisher_name()}}
\item Fuzzy matching: \code{\link[=find_journal]{find_journal()}}
}
}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

}
