% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.r
\name{find_journals}
\alias{find_journals}
\alias{journal_name}
\alias{journal_issn}
\alias{publisher_name}
\title{Find journal policy scores}
\usage{
journal_name(x, exact_match = TRUE, detail = level_detail$minimal)

journal_issn(x, exact_match = TRUE, detail = level_detail$minimal)

publisher_name(x, exact_match = TRUE, detail = level_detail$minimal)
}
\arguments{
\item{x}{character or numeric vector, can be length > 1}

\item{exact_match}{whether to only return exact matches, or use input as a sub-string
for grepping}

\item{detail}{level of detail to return, either "minimal" or NULL}
}
\description{
Find journals in the TOP dataset, through either the journal name,
ISSN, or publisher name.
}
